made weapon pickups generic
- deleted weapon shotgun pickup scene and script
- made a generic weapon pickup scene, and then created weapon pickup scenes for revolver and shotgun which uses that scene
