+++ Assets/Sprites/PlayerSprites/cat_revolver-1.png.import
@@ -1,0 +1,34 @@
+[remap]
+
+importer="texture"
+type="CompressedTexture2D"
+uid="uid://cywupqthyqppa"
+path="res://.godot/imported/cat_revolver-1.png-6a0f14749fb2cc955b2ac5a0c3fe51a8.ctex"
+metadata={
+"vram_texture": false
+}
+
+[deps]
+
+source_file="res://Assets/Sprites/PlayerSprites/cat_revolver-1.png"
+dest_files=["res://.godot/imported/cat_revolver-1.png-6a0f14749fb2cc955b2ac5a0c3fe51a8.ctex"]
+
+[params]
+
+compress/mode=0
+compress/high_quality=false
+compress/lossy_quality=0.7
+compress/hdr_compression=1
+compress/normal_map=0
+compress/channel_pack=0
+mipmaps/generate=false
+mipmaps/limit=-1
+roughness/mode=0
+roughness/src_normal=""
+process/fix_alpha_border=true
+process/premult_alpha=false
+process/normal_map_invert_y=false
+process/hdr_as_srgb=false
+process/hdr_clamp_exposure=false
+process/size_limit=0
+detect_3d/compress_to=1
