+++ Weapons/Pickups/Scenes/weapon_pickup_shotgun.tscn
@@ -1,0 +1,49 @@
+[gd_scene load_steps=8 format=3 uid="uid://xgbwu1var27k"]
+
+[ext_resource type="PackedScene" uid="uid://jrk5sgo60jhm" path="res://Weapons/Pickups/Scenes/weapon_pickups.tscn" id="1_8whpw"]
+[ext_resource type="Texture2D" uid="uid://8qngepld1gb2" path="res://Assets/Sprites/PlayerSprites/cat_shotgun-3.png" id="2_lt7pk"]
+
+[sub_resource type="AtlasTexture" id="AtlasTexture_0h2br"]
+atlas = ExtResource("2_lt7pk")
+region = Rect2(0, 32, 48, 32)
+
+[sub_resource type="AtlasTexture" id="AtlasTexture_nwlil"]
+atlas = ExtResource("2_lt7pk")
+region = Rect2(48, 32, 48, 32)
+
+[sub_resource type="AtlasTexture" id="AtlasTexture_frph6"]
+atlas = ExtResource("2_lt7pk")
+region = Rect2(96, 32, 48, 32)
+
+[sub_resource type="AtlasTexture" id="AtlasTexture_k55hr"]
+atlas = ExtResource("2_lt7pk")
+region = Rect2(144, 32, 48, 32)
+
+[sub_resource type="SpriteFrames" id="SpriteFrames_85v04"]
+animations = [{
+"frames": [{
+"duration": 1.0,
+"texture": SubResource("AtlasTexture_0h2br")
+}, {
+"duration": 1.0,
+"texture": SubResource("AtlasTexture_nwlil")
+}, {
+"duration": 1.0,
+"texture": SubResource("AtlasTexture_frph6")
+}, {
+"duration": 1.0,
+"texture": SubResource("AtlasTexture_k55hr")
+}],
+"loop": true,
+"name": &"idle",
+"speed": 9.0
+}]
+
+[node name="weapon_pickup_shotgun" type="Node2D"]
+
+[node name="weapon_pickups" parent="." instance=ExtResource("1_8whpw")]
+
+[node name="sprite" type="AnimatedSprite2D" parent="."]
+position = Vector2(0, 1)
+sprite_frames = SubResource("SpriteFrames_85v04")
+animation = &"idle"
