+++ Weapons/Pickups/Scenes/weapon_pickups.tscn
@@ -1,0 +1,16 @@
+[gd_scene load_steps=3 format=3 uid="uid://jrk5sgo60jhm"]
+
+[ext_resource type="Script" path="res://Weapons/Pickups/Scripts/WeaponPickups.cs" id="1_3d7k7"]
+
+[sub_resource type="CircleShape2D" id="CircleShape2D_m53o5"]
+radius = 11.0
+
+[node name="weapon_pickups" type="Area2D"]
+collision_layer = 4
+collision_mask = 2
+script = ExtResource("1_3d7k7")
+
+[node name="collider" type="CollisionShape2D" parent="."]
+position = Vector2(1, 7)
+shape = SubResource("CircleShape2D_m53o5")
+debug_color = Color(0.878209, 0.266446, 0.545585, 0.42)
